+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/ui/BaseSheetActivity.kt
@@ -1,10 +1,9 @@
 package com.stripe.android.paymentsheet.ui
 
-import android.os.Build
 import android.os.Bundle
 import androidx.activity.addCallback
+import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.WindowCompat
 import com.stripe.android.paymentsheet.LinkHandler
 import com.stripe.android.paymentsheet.viewmodels.BaseSheetViewModel
 import com.stripe.android.utils.fadeOut
@@ -20,13 +19,12 @@ internal abstract class BaseSheetActivity<ResultType> : AppCompatActivity() {
     abstract fun setActivityResult(result: ResultType)
 
     override fun onCreate(savedInstanceState: Bundle?) {
+        enableEdgeToEdge()
         super.onCreate(savedInstanceState)
 
         if (earlyExitDueToIllegalState) {
             return
         }
-
-        renderEdgeToEdge()
 
         onBackPressedDispatcher.addCallback {
             viewModel.handleBackPressed()
@@ -36,13 +34,5 @@ internal abstract class BaseSheetActivity<ResultType> : AppCompatActivity() {
     override fun finish() {
         super.finish()
         fadeOut()
-    }
-
-    private fun renderEdgeToEdge() {
-        if (Build.VERSION.SDK_INT < 30) {
-            return
-        }
-
-        WindowCompat.setDecorFitsSystemWindows(window, false)
     }
 }